Allow custom update hooks
I know the prefered way is using web hooks.
But we also have some bash scripts which we realy need. For example we
want things like rsync to trigger, doing this using webhooks would
require me to write an HTTP app for this.

The same is true for post-receive hooks, we would like to force server
side syntax checking and style checks.

I understand that this is not maintained by gitlab therefore call all
scripts located in the update.d