Remove recursive commands from bin/install
It is not clear what we need these for anymore. Because these commands
would recurse once through all existing files in all Git repositories,
and then another time through all directories these commands could
take very long on a server with a lot of repositories.

We keep the 'chmod' on config.repos_path, to deny world access to all
files and directories beneath it.