ARGF does more than we want
ARGF is a virtual IO stream that treats the elements of ARGV as paths
to files to be opened. In practice, git-receive-pack will probably
never pass arguments to pre-receive and post-receive because it is not
supposed to (according to 'man 5 githooks'). In case it does, however
our hooks should not have the surprise behavior of treating those
arguments as files to be opened.