Get secret from config
This adds parsing if the `secret_file` from YAML. And reads the
contents of the configured file.

If no file is configured, we fall back to the `.gitlab_shell_secret`
in the root dir of the configuration.

If the configured path does not start with a `/` we start looking for
the file relative to the root dir of the configuration.