Decouple bin/compile and bin/install from lib/
This commit duplicates a small amount of Ruby code that is run at build
or install time. Over time, it would be nice to remove the dependency
on Ruby for these steps, but the focus right now is on run-time.

We'll be removing lib/* from gitlab-shell very soon, and these commands
need to keep working when we do.