commands: pass through GIT_PROTOCOL envvar provided by clients
Both git-upload-pack and git-receive-pack services inspect the
GIT_PROTOCOL environment transferred via SSH in order to decide which
protocols are supported by a given client. Currently, we don't use the
environment variable at all, though, but instead forward the GitProtocol
field of the access verification response.

Improve this by passing on the GIT_PROTOCOL environment variable
provided by the client as-is.