Stop changing directory to the filesystem root
This behaviour dates from when Gitaly RPCs were executed in Ruby by a
Go subprocess. It's not needed for gitlab-shell now that it's in Go,
and it's a very strange thing for gitlab-sshd. Best just to remove it.

If we wanted to retain this behaviour, we could have an `os.Chdir` call
in the gitlab-shell binary, but I just don't think it's needed.

Changelog: fixed