Fix usage of out-of-date Gitaly images
Our CI jobs and docker-compose pull in the "latest" tag of Gitaly. As it
turns out though, "latest" is pointing to Gitaly v13.3.0-rc5, which is
definitely not the latest versionat this point in time. This is because
CNG was converted to not use the "latest" tag anymore, but instead to
use a tag called "master" in gitlab-org/build/CNG!519.

Fix this by using the new "master" tag instead.