Update Gitaly dependency to v14.9.0-rc1
This commit updates the Gitaly dependency's version to v14.9.0-rc1.
This version contains a fix for connection hangs when gRPC closes
an unhealthy connection to Gitaly. This bug likely never affected
gitlab-shell due to it not reusing connections to Gitaly. To avoid
problems in the future if that is ever done, let's update the
dependency.