Default to info level for an empty log-level
I'd assumed that the `omitempty` directive for LogLevel in
internal/config/config.go would get us this behaviour. If it did, we
wouldn't have had to specify the default twice. Unfortunately, it
doesn't, which is to say that given a config file like:

```
log_level:
```

The default *is* overridden by the empty string.

It's an easy enough fix.

Changelog: fixed