Wait until all Gitaly sessions are executed
If they haven't been executed within a timeout, we unblock the
execution.

Issue: https://gitlab.com/gitlab-org/gitlab-shell/-/issues/563

We have an issue when a lot of git clones operations get canceled.
It was assumed that some clients close the connection just after
all the data has been received from Git server. If there was a
network delay and gitlab-sshd hadn't managed to gracefully close
the connection, context get canceled and Gitaly cancels the
execution and returns the error.

Let's wait for a perion to allow Gitaly to gracefully complete the
operation