Extend shell logs with an extra metadata
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/442907

**Problem**

We discovered that reliance on project and namespace paths is not
sufficient. User can rename the group or project and that might lead to
problems with identifying the correct project/namespace.

**Solution**

Instead of project/namespace path we should use their IDs.

Changelog: added