Fix race condition by using io.NopCloser wrapper
The issue was that the Gitaly client expects an io.ReadCloser interface
for stdin, but strings.NewReader only implements io.Reader. This can
cause race conditions when the client tries to close the stream.

Wrapping with io.NopCloser provides the expected interface and prevents
the "stdin send error: EOF" race condition.