Attempts to migrate gitlab-sshd client over to v2 labkit logger
Now that we've validated that the v2 LabKit logs successfully flow
through our system into ElasticSearch, we can now proceed with
confidence and make broader changes to this system to migrate it
over.

This commit focuses on migrating the gitlab-sshd binary over fully
as this is the easier of the 2 logging binaries.